Added More JS Functionality
Continuing to expand the functionality of <main.js> and the elements to
support it in <index.html>. Specifically in this commit, the storage of
all messages the user has submitted in an Array that is displayed under
the most current message. Also added some semicolons throughout at the
end of jQuery functions.